refactor(MainViewModel): Move `markAsSeen()` to `SharedViewModelUtils`
