+++ app/src/main/java/com/infomaniak/mail/ui/MainViewModel.kt
@@ -25,7 +25,10 @@ import com.infomaniak.mail.R
 import com.infomaniak.mail.data.LocalSettings
 import com.infomaniak.mail.data.api.ApiRepository
 import com.infomaniak.mail.data.cache.RealmDatabase
-import com.infomaniak.mail.data.cache.mailboxContent.*
+import com.infomaniak.mail.data.cache.mailboxContent.FolderController
+import com.infomaniak.mail.data.cache.mailboxContent.MessageController
+import com.infomaniak.mail.data.cache.mailboxContent.SignatureController
+import com.infomaniak.mail.data.cache.mailboxContent.ThreadController
 import com.infomaniak.mail.data.cache.mailboxInfo.MailboxController
 import com.infomaniak.mail.data.cache.mailboxInfo.QuotasController
 import com.infomaniak.mail.data.cache.userInfo.AddressBookController
@@ -41,6 +44,7 @@ import com.infomaniak.mail.data.models.thread.Thread.ThreadFilter
 import com.infomaniak.mail.utils.AccountUtils
 import com.infomaniak.mail.utils.ContactUtils.getPhoneContacts
 import com.infomaniak.mail.utils.ContactUtils.mergeApiContactsIntoPhoneContacts
+import com.infomaniak.mail.utils.SharedViewModelUtils.markAsSeen
 import com.infomaniak.mail.utils.Utils.formatFoldersListWithAllChildren
 import com.infomaniak.mail.workers.DraftsActionsWorker
 import io.realm.kotlin.ext.copyFromRealm
@@ -302,53 +306,6 @@ class MainViewModel(application: Application) : AndroidViewModel(application) {
         refreshThreads()
     }
 
-    //region Open Thread
-    fun openThread(threadUid: String) = liveData(Dispatchers.IO) {
-
-        val mailbox = MailboxController.getMailbox(AccountUtils.currentUserId, AccountUtils.currentMailboxId) ?: run {
-            emit(null)
-            return@liveData
-        }
-
-        val thread = ThreadController.getThread(threadUid) ?: run {
-            emit(null)
-            return@liveData
-        }
-
-        val expandedList = thread.messages.mapIndexed { index, message ->
-            !message.isDraft && (!message.seen || index == thread.messages.lastIndex)
-        }
-        emit(expandedList)
-
-        fetchIncompleteMessages(thread)
-
-        if (thread.unseenMessagesCount > 0) {
-            markAsSeen(thread, mailbox.uuid)
-            MessageController.fetchCurrentFolderMessages(mailbox, thread.folderId, localSettings.threadMode)
-        }
-    }
-
-    private fun fetchIncompleteMessages(thread: Thread) {
-        RealmDatabase.mailboxContent().writeBlocking {
-            thread.messages.forEach { localMessage ->
-                if (!localMessage.fullyDownloaded) {
-                    ApiRepository.getMessage(localMessage.resource).data?.also {
-                        it.messageIds = localMessage.messageIds
-
-                        // If we've already got this Message's Draft beforehand, we need to save
-                        // its `draftLocalUuid`, otherwise we'll lose the link between them.
-                        if (it.isDraft) it.draftLocalUuid = DraftController.getDraftByMessageUid(it.uid, realm = this)?.localUuid
-
-                        it.fullyDownloaded = true
-
-                        MessageController.upsertMessage(it, realm = this)
-                    }
-                }
-            }
-        }
-    }
-    //endregion
-
     //region Seen status
     fun toggleSeenStatus(threadUid: String) = viewModelScope.launch(Dispatchers.IO) {
         val mailboxUuid = currentMailbox.value?.uuid ?: return@launch
@@ -362,12 +319,6 @@ class MainViewModel(application: Application) : AndroidViewModel(application) {
         val uids = ThreadController.getThreadLastMessageUids(thread)
 
         ApiRepository.markMessagesAsUnseen(mailboxUuid, uids)
-    }
-
-    private fun markAsSeen(thread: Thread, mailboxUuid: String) {
-        val uids = ThreadController.getThreadUnseenMessagesUids(thread)
-
-        ApiRepository.markMessagesAsSeen(mailboxUuid, uids)
     }
     //endregion
