+++ app/src/main/java/com/infomaniak/mail/ui/main/thread/ThreadFragment.kt
@@ -64,7 +64,7 @@ class ThreadFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         observeThreadLive()
-        mainViewModel.openThread(navigationArgs.threadUid).observe(viewLifecycleOwner) { expandedList ->
+        threadViewModel.openThread(navigationArgs.threadUid).observe(viewLifecycleOwner) { expandedList ->
             if (expandedList == null) {
                 findNavController().popBackStack()
             } else {
