+++ app/src/main/java/com/infomaniak/mail/utils/SharedViewModelUtils.kt
@@ -1,0 +1,30 @@
+/*
+ * Infomaniak kMail - Android
+ * Copyright (C) 2022 Infomaniak Network SA
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+package com.infomaniak.mail.utils
+
+import com.infomaniak.mail.data.api.ApiRepository
+import com.infomaniak.mail.data.cache.mailboxContent.ThreadController
+import com.infomaniak.mail.data.models.thread.Thread
+
+object SharedViewModelUtils {
+    fun markAsSeen(thread: Thread, mailboxUuid: String) {
+        val uids = ThreadController.getThreadUnseenMessagesUids(thread)
+
+        ApiRepository.markMessagesAsSeen(mailboxUuid, uids)
+    }
+}
